backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.25
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.2.0
httpx>=0.28.0
feedparser>=6.0.10
python-multipart>=0.0.18
aiosqlite>=0.20.0
tenacity>=8.2.0
yt-dlp>=2025.0.0
# Issue #12: Scheduled Backups
APScheduler>=3.10.0
aiofiles>=24.1.0
# Issue #41: Optional Authentication (JWT only - hashing uses built-in hashlib)
PyJWT>=2.10.0